Move main and polyfill to src
