+++ bitwarden_license/bit-web/src/main.ts
@@ -5,10 +5,10 @@ import "bootstrap";
 import "jquery";
 import "popper.js";
 
-require("../scss/styles.scss");
-require("../scss/tailwind.css");
+require("src/scss/styles.scss");
+require("src/scss/tailwind.css");
 
-import { AppModule } from "./app.module";
+import { AppModule } from "./app/app.module";
 
 if (process.env.NODE_ENV === "production") {
   enableProdMode();
